+++ test/Integration/CosmosDB.integration.Tests.ps1
@@ -68,6 +68,7 @@ $script:testDocumentUTF8UpdateBody = @"
 $script:testAttachmentId = 'testAttachment'
 $script:testAttachmentContentType = 'image/jpg'
 $script:testAttachmentMedia = 'www.bing.com'
+$script:testAttachmentMediaUpdated = 'www.contoso.com'
 $script:testStoredProcedureId = 'testStoredProcedure'
 $script:testStoredProcedureBody = @'
 function () {
@@ -794,6 +795,26 @@ Describe 'Cosmos DB Module' -Tag 'Integration' {
         }
     }
 
+    Context 'When updating an attachment from the document in a collection' {
+        It 'Should not throw an exception' {
+            $script:result = Set-CosmosDbAttachment `
+                -Context $script:testContext `
+                -CollectionId $script:testCollection `
+                -DocumentId $script:testDocumentId `
+                -Id $script:testAttachmentId `
+                -ContentType $script:testAttachmentContentType `
+                -Media $script:testAttachmentMediaUpdated `
+                -Verbose
+        }
+
+        It 'Should return expected object' {
+            Test-GenericResult -GenericResult $script:result
+            $script:result.Id | Should -Be $script:testAttachmentId
+            $script:result.ContentType | Should -Be $script:testAttachmentContentType
+            $script:result.Media | Should -Be $script:testAttachmentMediaUpdated
+        }
+    }
+
     Context 'When adding a read document permission for a user' {
         It 'Should not throw an exception' {
             $script:documentResourcePath = Get-CosmosDbDocumentResourcePath `
@@ -1185,6 +1206,79 @@ Describe 'Cosmos DB Module' -Tag 'Integration' {
             $script:result.Id | Should -Be $script:testDocumentId
             $script:result.Content | Should -Be 'Some string'
             $script:result.More | Should -Be 'Some other string'
+        }
+    }
+
+    Context 'When adding an attachment to the document in a collection with a partition key' {
+        It 'Should not throw an exception' {
+            $script:result = New-CosmosDbAttachment `
+                -Context $script:testContext `
+                -CollectionId $script:testCollection `
+                -DocumentId $script:testDocumentId `
+                -Id $script:testAttachmentId `
+                -PartitionKey $script:testDocumentId `
+                -ContentType $script:testAttachmentContentType `
+                -Media $script:testAttachmentMedia `
+                -Verbose
+        }
+
+        It 'Should return expected object' {
+            Test-GenericResult -GenericResult $script:result
+            $script:result.Id | Should -Be $script:testAttachmentId
+            $script:result.ContentType | Should -Be $script:testAttachmentContentType
+            $script:result.Media | Should -Be $script:testAttachmentMedia
+        }
+    }
+
+    Context 'When getting an attachment from the document in a collection with a partition key' {
+        It 'Should not throw an exception' {
+            $script:result = Get-CosmosDbAttachment `
+                -Context $script:testContext `
+                -CollectionId $script:testCollection `
+                -DocumentId $script:testDocumentId `
+                -Id $script:testAttachmentId `
+                -PartitionKey $script:testDocumentId `
+                -Verbose
+        }
+
+        It 'Should return expected object' {
+            Test-GenericResult -GenericResult $script:result
+            $script:result.Id | Should -Be $script:testAttachmentId
+            $script:result.ContentType | Should -Be $script:testAttachmentContentType
+            $script:result.Media | Should -Be $script:testAttachmentMedia
+        }
+    }
+
+    Context 'When updating an attachment from the document in a collection with a partition key' {
+        It 'Should not throw an exception' {
+            $script:result = Set-CosmosDbAttachment `
+                -Context $script:testContext `
+                -CollectionId $script:testCollection `
+                -DocumentId $script:testDocumentId `
+                -Id $script:testAttachmentId `
+                -PartitionKey $script:testDocumentId `
+                -ContentType $script:testAttachmentContentType `
+                -Media $script:testAttachmentMediaUpdated `
+                -Verbose
+        }
+
+        It 'Should return expected object' {
+            Test-GenericResult -GenericResult $script:result
+            $script:result.Id | Should -Be $script:testAttachmentId
+            $script:result.ContentType | Should -Be $script:testAttachmentContentType
+            $script:result.Media | Should -Be $script:testAttachmentMediaUpdated
+        }
+    }
+
+    Context 'When removing an attachment from the document in a collection with a partition key' {
+        It 'Should not throw an exception' {
+            $script:result = Remove-CosmosDbAttachment `
+                -Context $script:testContext `
+                -CollectionId $script:testCollection `
+                -DocumentId $script:testDocumentId `
+                -Id $script:testAttachmentId `
+                -PartitionKey $script:testDocumentId `
+                -Verbose
         }
     }